DATA: revert to float32

+++ edge_tpu_video_style/preprocessing/read_mpi.py
@@ -64,24 +64,24 @@ class MPIDataSet:
                 flow = readFlow(str(flowpath.joinpath(f"frame_{num1}.flo")))
 
                 img1 = (
-                    tf.convert_to_tensor(np.array(img1), dtype=tf.float16)
+                    tf.convert_to_tensor(np.array(img1), dtype=tf.float32)
                     / float_conv_factor
                 )
                 img1 = tf.transpose(img1, (1, 0, 2))
                 img2 = (
-                    tf.convert_to_tensor(np.array(img2), dtype=tf.float16)
+                    tf.convert_to_tensor(np.array(img2), dtype=tf.float32)
                     / float_conv_factor
                 )
                 img2 = tf.transpose(img2, (1, 0, 2))
                 h, w, c = flow.shape
 
-                flow = tf.convert_to_tensor(np.array(flow), dtype=tf.float16)
+                flow = tf.convert_to_tensor(np.array(flow), dtype=tf.float32)
                 flow = tf.transpose(flow, (1, 0, 2))
                 flow = tf.image.resize(
                     images=flow, size=(self.args.width, self.args.height)
                 )
                 multiplier = tf.constant(
-                    [flow.shape[0] / w, flow.shape[1] / h], dtype=tf.float16
+                    [flow.shape[0] / w, flow.shape[1] / h], dtype=tf.float32
                 )
                 flow *= multiplier
 
@@ -92,7 +92,7 @@ class MPIDataSet:
                 mask = 1 - mask
                 mask = np.where((mask < 0.99), 0.0, 1.0)
                 # now convert to tensor
-                mask = tf.convert_to_tensor(np.array(mask), dtype=tf.float16)
+                mask = tf.convert_to_tensor(np.array(mask), dtype=tf.float32)
                 mask = tf.expand_dims(mask, axis=0)
                 mask = tf.transpose(mask, (2, 1, 0))
                 break
@@ -116,5 +116,5 @@ def read_style_image(args):
     style_img = style_img.resize((args.width, args.height), Image.BILINEAR)
     style_img = tf.convert_to_tensor(np.array(style_img))
     style_img = tf.transpose(style_img, (1, 0, 2))
-    style_img = tf.cast(tf.expand_dims(style_img / 255, axis=0), dtype=tf.float16)
+    style_img = tf.cast(tf.expand_dims(style_img / 255, axis=0), dtype=tf.float32)
     return style_img
